+++ parkflow-back-end/pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
-	<artifactId>demo1</artifactId>
+	<artifactId>parkflow</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>demo1</name>
 	<description>Demo project for Spring Boot</description>
