+++ parkflow-back-end/src/main/java/com/example/parkflow/ParkflowApplication.java
@@ -1,13 +1,13 @@
-package com.example.demo1;
+package com.example.parkflow;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Demo1Application {
+public class ParkflowApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(Demo1Application.class, args);
+		SpringApplication.run(ParkflowApplication.class, args);
 	}
 
 }
